Add scale-6x-stable; Add 'master' resources
With the addition of the 6X_STABLE branch, we must ensure that there is
consistency between the naming of jobs and the branches they track. The
'master' jobs now point to a binary and src that correspond to the
master branch of GPDB, while the newly-added 6X jobs use resources
pulled from the 6X_STABLE branch.

Authored-by: Trevor Yacovone <tyacovone@pivotal.io>